+++ app/nodeos_mindreader/app.go
@@ -55,12 +55,21 @@ type Config struct {
 	NodeosExtraArgs []string
 
 	// Common Flags
-	BackupStoreURL      string
-	BackupTag           string
+	BackupStoreURL string
+
+	// Backup Flags
+	BackupTag        string
+	AutoBackupModulo int
+	AutoBackupPeriod time.Duration
+
+	// Snapshot Flags
+	AutoSnapshotModulo int
+	AutoSnapshotPeriod time.Duration
+
 	BootstrapDataURL    string
 	DebugDeepMind       bool
 	LogToZap            bool
-	AutoRestoreLatest   bool
+	AutoRestoreSource   string
 	RestoreBackupName   string
 	RestoreSnapshotName string
 	SnapshotStoreURL    string
@@ -137,10 +146,6 @@ func (a *App) Run() error {
 	dmetrics.Register(metrics.NodeosMetricset)
 	dmetrics.Register(metrics.Metricset)
 
-	if a.Config.StartFailureHandlerFunc != nil {
-		chainSuperviser.Superviser.RegisterStartFailureHandler(a.Config.StartFailureHandlerFunc)
-	}
-
 	var p *profiler.Profiler
 	if !a.Config.DisableProfiler {
 		p = profiler.MaybeNew()
@@ -150,7 +155,7 @@ func (a *App) Run() error {
 		BootstrapDataURL:           a.Config.BootstrapDataURL,
 		BackupTag:                  a.Config.BackupTag,
 		BackupStoreURL:             a.Config.BackupStoreURL,
-		AutoRestoreLatest:          a.Config.AutoRestoreLatest,
+		AutoRestoreSource:          a.Config.AutoRestoreSource,
 		ShutdownDelay:              a.Config.ShutdownDelay,
 		RestoreBackupName:          a.Config.RestoreBackupName,
 		RestoreSnapshotName:        a.Config.RestoreSnapshotName,
@@ -159,9 +164,16 @@ func (a *App) Run() error {
 		Profiler:                   p,
 		ReadyFunc:                  a.ReadyFunc,
 	})
+	//if a.Config.StartFailureHandlerFunc != nil {
+	//	chainOperator.RegisterStartFailureHandler(a.Config.StartFailureHandlerFunc)
+	//}
+
 	if err != nil {
 		return fmt.Errorf("unable to create chain operator: %w", err)
 	}
+
+	chainOperator.ConfigureAutoBackup(a.Config.AutoBackupPeriod, a.Config.AutoBackupModulo)
+	chainOperator.ConfigureAutoSnapshot(a.Config.AutoSnapshotPeriod, a.Config.AutoSnapshotModulo)
 
 	gs := dgrpc.NewServer(dgrpc.WithLogger(zlog))